track info object does not allow attributes to be set to empty string
players providing incomplete metadata information will have
the missing properties set as 'unknown'

+++ builder/java-tools/music-player/src/main/java/com/conky/musicplayer/AvailabilityHandler.java
@@ -51,9 +51,12 @@ public class AvailabilityHandler extends AbstractSignalHandlerBase<DBus.NameOwne
 
         if (!signal.oldOwner.isEmpty() && signal.newOwner.isEmpty()) {
             String playerName = signal.name.substring(signal.name.lastIndexOf('.') + 1);
-            logger.debug("'{}' music player has shutdown", playerName);
-            playerDatabase.removePlayer(signal.oldOwner);
-            writer.writePlayerState(playerDatabase.getActivePlayer());
+
+            if (playerDatabase.isMusicPlayer(playerName)) {
+                logger.info("the '{}' music player is no longer running", playerName);
+                playerDatabase.removePlayer(signal.oldOwner);
+                writer.writePlayerState(playerDatabase.getActivePlayer());
+            }
         }
     }
 }
